+++ devito/dle/backends/basic.py
@@ -11,7 +11,7 @@ from devito.cgen_utils import ccode
 from devito.dse import as_symbol
 from devito.dle import retrieve_iteration_tree, filter_iterations
 from devito.dle.backends import AbstractRewriter, dle_pass, complang_ALL
-from devito.interfaces import Scalar
+from devito.types import Scalar
 from devito.nodes import (Denormals, Expression, Call, Callable, List,
                           UnboundedIndex)
 from devito.tools import filter_sorted, flatten
