+++ devito/dse/clusterizer.py
@@ -3,7 +3,7 @@ from collections import OrderedDict, namedtuple
 from devito.dse.graph import temporaries_graph
 from devito.dse.manipulation import xreplace_indices
 
-from devito.interfaces import Scalar
+from devito.types import Scalar
 from devito.stencil import Stencil
 from devito.tools import as_tuple
