+++ devito/function.py
@@ -10,7 +10,7 @@ from devito.memory import CMemory, first_touch
 from devito.cgen_utils import INT, FLOAT
 from devito.dimension import d, p, t, time, x, y, z
 from devito.arguments import ConstantArgProvider, TensorFunctionArgProvider
-from devito.interfaces import SymbolicFunction, AbstractSymbol
+from devito.types import SymbolicFunction, AbstractSymbol
 from devito.finite_difference import (centered, cross_derivative,
                                       first_derivative, left, right,
                                       second_derivative, generic_derivative,
