+++ devito/nodes.py
@@ -10,7 +10,7 @@ from sympy import Eq
 
 from devito.cgen_utils import ccode
 from devito.dse import as_symbol, retrieve_terminals
-from devito.interfaces import Indexed, Symbol
+from devito.types import Indexed, Symbol
 from devito.stencil import Stencil
 from devito.tools import as_tuple, filter_ordered, flatten, is_integer
 from devito.arguments import ArgumentProvider, Argument
