+++ devito/types.py
